Refact: moved common types to types.go
